add/9-dummy data

+++ models/custom3.js
@@ -21,7 +21,7 @@ module.exports = (sequelize, DataTypes) => {
     }
   };
   custom3.init({
-    value: DataTypes.INTEGER,
+    value: DataTypes.FLOAT,
     user_id: DataTypes.INTEGER,
     scheldule: DataTypes.DATEONLY
   }, {
